Fix paging problems; refactoring

+++ kernel/cpu/gdt.hpp
@@ -106,7 +106,7 @@ struct tss final {
     tss() = default;
 
     tss(void *kernel_stack) {
-        utils::fill(io_bitmap, io_bitmap + 128, static_cast<uint8_t>(0));
+        utils::fill(io_bitmap, io_bitmap + 128, 0u);
         iomap_offset = 104;
         ss0 = segment::kernel_ds;
         esp0 = reinterpret_cast<uint32_t>(kernel_stack);
